USF-HELPER/requirements.txt
pytz
discord.py
selenium
webdriver_manager